Refine CSS targeting of today's background color
The old CSS selector for setting the background color of the
highlighted day (today) also affected `<td>` elements within
`.fc-content-skeleton`.

The FullCalendar CSS gives `.fc-content-skeleton` a transparent
background and border, so that the underlying background and border of
`.fc-bg` are shown.

Overriding of the `background-color` of `.fc-content-skeleton` by
ownCloud inline CSS can, under certain circumstances, lead to that
background color extending under the (transparent) border of
`.fc-content-skeleton`, thereby overlapping the border of `.fc-bg`
below.

This change prevents that by specifically targeting the `.fc-bg` class.