Fix blank page on browsers without support for ResizeObserver
ResizeObserver was used unconditionally, which caused the Calendar to
fail to load and show a blank page on some older browsers without
support for it. Now ResizeObserver is used only when supported.

ResizeObserver ensures that the calendar grid is properly sized when the
navigation bar and sidebar are opened and closed. It also ensures that
the text areas are properly sized when shown in the sidebar.

Given that they are mostly "cosmetic" issues and that they should not
affect many users no fallback was added for those browsers without
support for ResizeObserver.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>