Use npm ci to install front-end dependencies in dev
The ci command is meant for any situation where you want to make sure
you're doing a clean install of your dependencies.

It is significantly faster than npm install. If a node_modules is
already present, it will be automatically removed before npm ci begins
its install.

Matches the GitHub actions to build the package.

The clean-dev rule is now unused and can be removed.

https://docs.npmjs.com/cli/v7/commands/npm-ci
Signed-off-by: François Freitag <mail@franek.fr>