Fix sorting of free/busy resources
Fullcalendar sorts resources by their ID. In our app that is the
attendee's email address. For the fake blocking event that shows the
current time slot we need to know the first and last resource to show a
special border on top/bottom. In some scenarios that didn't work before,
now it seems more reliable.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>