Fix free/busy blocking event background and border
The blocking event (the slot for the selected time frame of the event)
is supposed to show with red border and no background. For some reason
this styling broke (possibly through a fullcalendar upgrade) and
therefore needed some stricter styling rules that turn off FC's default
background and reset the opacity to make the border stand out.

This brings the design back to the original design as done by Georg.
Ref https://github.com/nextcloud/calendar/pull/1731

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>