Add setting for sync timeout
The setting is a string that can be formatted as anything that
[moment.duration](https://momentjs.com/docs/#/durations/creating/)
can parse:
* 'hh:mm:ss'
* 'PT1M' (ISO 8601)

Any invalid setting or setting to less than a second
will lead to the calendar updates being disabled.

The setting is not exposed in the UI.
It's meant to be configured by server admins
to be able to adjust the timeout and prevent an overly high load.

Signed-off-by: Azul <azul@riseup.net>